:sparkles: Add functionality to load investment details after submission
This commit adds functionality to load investment details after submitting a new investment or updating an existing one. The `InvestmentSubmitted` state is now handled in the `InvestmentPage` widget, where it triggers the loading of the submitted investment details. Additionally, the `Navigator` now pops the page with the updated investment when the delete operation is successful.

:recycle: refactor(investments): Remove redundant error handling code

In the `InvestmentsRepositoryImpl`, the error handling code for parsing error messages from Dio responses has been simplified. Now, the `responseData` is checked directly from the error object without unnecessary conditional statements.

:lipstick: style(investments): Update button text in delete confirmation

The button text for deleting an investment has been updated to remove the period at the end, providing a cleaner and more consistent user interface.

:bug: fix(investments): Fix error handling in InvestmentsRepositoryImpl

Fixed a bug in the `InvestmentsRepositoryImpl` where error handling was not correctly checking for the response data type before attempting to parse it. Now, the error is properly handled to prevent potential crashes.

:art: style(ui): Update UI consistency in investment details screen

Adjusted the UI layout and styling in the `InvestmentDetailsPage` to improve consistency and readability. Removed unnecessary loading indicators and enhanced the display of investment details.

:hammer: chore(investments): Clean up redundant code and imports

Removed redundant code and unused imports from various files to improve code cleanliness and maintainability. Updated constants and refactored code for better organization.